requirements_file.txt
# Core ML and Data Processing
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
lightgbm>=3.3.0
joblib>=1.2.0

# Database and Storage
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# API and WebSocket
requests>=2.28.0
websockets>=10.0
dhanhq>=1.3.0

# Async and Threading
asyncio
aiofiles>=0.8.0

# Utilities
python-dotenv>=0.19.0
argparse
uuid
json5>=0.9.0

# Statistical Analysis
scipy>=1.9.0
statsmodels>=0.13.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
isort>=5.10.0
flake8>=5.0.0
mypy>=0.991

# Optional: For enhanced performance
numba>=0.56.0
bottleneck>=1.3.0

# Optional: For visualization and monitoring
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0